requirements.txt
Flask
Flask-Mongo
Python-dotenv